Review comments
Put es_process_found back in and checking inside the loop

+++ cs-osd/cs-osd.py
@@ -415,14 +415,17 @@ def checkTemperature():
 def doShutdown():
     proc = subprocess.Popen(['ps', '-C', 'emulationstation', '-o', 'pid,args'], stdout = subprocess.PIPE)
     out, err = proc.communicate()
+    
+    es_process_found = False
     for line in out.splitlines():
-        path = line.split(None, 1)[1]
-        if path.startswith('/opt/retropie/supplementary/emulationstation/emulationstation'):
-            es_process_found = True
-            open('/tmp/es-shutdown', 'a').close()
-            pid = int(line.split(None, 1)[0])
-            os.kill(pid, signal.SIGTERM)
-            time.sleep(5)
+        if not es_process_found:
+            path = line.split(None, 1)[1]
+            if path.startswith('/opt/retropie/supplementary/emulationstation/emulationstation'):
+                es_process_found = True
+                open('/tmp/es-shutdown', 'a').close()
+                pid = int(line.split(None, 1)[0])
+                os.kill(pid, signal.SIGTERM)
+                time.sleep(5)
 
     os.system("sudo shutdown -h now")
